requirements.txt
streamlit
numpy
pandas
scipy
joblib
scikit-learn
threadpoolctl
xgboost
lightgbm
catboost
matplotlib
openpyxl
openai
faiss-cpu
pymupdf